Rename `Bits` type and more instances.
